closing-issues-checker/requirements.txt
PyGithub==1.43.3
emoji==0.5.1
pyyaml==3.13
